Refactor common code for ReadBuilder for both non-vectorized and vectorized read
